Update Figs. 6 and 10, and update caption for Fig. 5.
Figs. 6 and 10 are formatted in the style of NMRlipids II (with scripts to make the new figures included in the Figs/scripts).

In Fig 5 caption, the M and F are now formatted just like in NMRlipids I. Alternate formatting could be {\textbf{\footnotesize F}} which makes the font bold and a tiny bit smaller.