Fixes for issues #38 and #78
Issue #38:
There was no source file for Figure 5 so the edits are done manually using Adobe Acrobat. May need a check that it is still ok.

Issue #78:
Could not run pdfcairo on my MacOS so only gnuplot .gpl file here.
Depending on OS there may be extra '\':s in the pdf figure, a fix is to replace all @{\"} with @{"}, running gnuplot again and then it should work.